requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5
